combined both SCVector and SRVector into the same file
