+++ src/lib.rs
@@ -25,9 +25,7 @@ pub use matrix::matrix_op::lu::{LUError, SLU};
 pub use matrix::matrix_op::qr::{QRError, SQR};
 pub use matrix::matrix_op::schur::{SchurError, SSchur};
 pub use matrix::matrix_op::eig::{EigError, SEig};
-pub use vector::scvector::{SCVector};
-pub use vector::srvector::{SRVector};
-// pub use matrix::addop::BinopExpr;
+pub use vector::svector::{SCVector, SRVector};
 
 
 //-------------------------------------------------------------------------------------------------
