+++ src/vector/dvector.rs
@@ -1,0 +1,16 @@
+//! Short Description of module
+//!
+//! Longer description of module
+//--------------------------------------------------------------------------------------------------
+
+//{{{ crate imports 
+use crate::matrix::dmatrix::DMatrix;
+//}}}
+//{{{ std imports 
+//}}}
+//{{{ dep imports 
+//}}}
+//--------------------------------------------------------------------------------------------------
+
+type DRVector<T> = DMatrix<T>;
+type DCVector<T> = DMatrix<T>;
